+++ venue/src/main/java/covia/adapter/agent/GoalTreeAdapter.java
@@ -16,7 +16,6 @@ import convex.core.data.Vectors;
 import convex.core.data.prim.CVMBool;
 import convex.core.data.prim.CVMLong;
 import convex.core.lang.RT;
-import covia.adapter.ContextBuilder;
 import covia.api.Fields;
 import covia.grid.Job;
 import covia.venue.AgentState;
