+++ venue/src/test/java/covia/adapter/AgentAdapterTest.java
@@ -18,6 +18,7 @@ import convex.core.data.Vectors;
 import convex.core.data.prim.CVMBool;
 import convex.core.data.prim.CVMLong;
 import convex.core.lang.RT;
+import covia.adapter.agent.ContextBuilder;
 import covia.api.Fields;
 import covia.grid.Job;
 import covia.grid.Status;
