+++ venue/src/test/java/covia/adapter/agent/ContextLoaderTest.java
@@ -1,4 +1,4 @@
-package covia.adapter;
+package covia.adapter.agent;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -17,6 +17,7 @@ import convex.core.data.Vectors;
 import convex.core.data.prim.CVMBool;
 import convex.core.data.util.CellExplorer;
 import convex.core.lang.RT;
+import covia.adapter.agent.ContextLoader;
 import covia.api.Fields;
 import covia.grid.Job;
 import covia.venue.Engine;
